Add oatpp to .travis.yml file
