+++ frameworks/C++/oatpp/README.md
@@ -1,0 +1,23 @@
+# Oat++ (AKA oatpp) Benchmarking Test
+
+- [Oat++ Github Repository](https://github.com/oatpp/oatpp)
+- [Oat++ Website](https://oatpp.io/)
+ 
+
+## Source for Tests
+
+### oatpp-async
+
+- [src-async](src-async/) - source code for oatpp test project implemented using oatpp async API.
+- [src-async/src/controller/MyController.hpp](src-async/src/controller/MyController.hpp) - Controller for `JSON`, `PLAINTEXT` endpoints.
+
+#### Test URLs
+
+- `JSON` - [http://127.0.0.1:8000/json](http://127.0.0.1:8000/json)
+- `PLAINTEXT` - [http://127.0.0.1:8000/plaintext](http://127.0.0.1:8000/plaintext) 
+
+## Contacts
+
+- Leonid Stryzhevskyi (Creator)
+   - Email: [leonid@oatpp.io](mailto:leonid@oatpp.io)
+   - Github: [lganzzzo](https://github.com/lganzzzo)
