+++ frameworks/C++/oatpp/src-async/src/controller/MyController.hpp
@@ -4,11 +4,11 @@
 
 #include "dto/DTOs.hpp"
 
+#include "Utils.hpp"
+
 #include "oatpp/web/server/api/ApiController.hpp"
 #include "oatpp/core/macro/codegen.hpp"
 #include "oatpp/core/macro/component.hpp"
-
-#include <chrono>
 
 /**
  * Sample Api Controller.
@@ -31,32 +31,6 @@ public:
  */
 #include OATPP_CODEGEN_BEGIN(ApiController)
 
-  static oatpp::String renderTime() {
-
-    static thread_local time_t lastSecond = 0;
-    static thread_local oatpp::String rederedTime;
-    static int count = 0;
-
-    auto time = std::chrono::system_clock::now().time_since_epoch();
-    time_t seconds = (time_t)std::chrono::duration_cast<std::chrono::seconds>(time).count();
-
-    if(seconds != lastSecond) {
-
-      lastSecond = seconds;
-
-      struct tm now;
-      gmtime_r(&seconds, &now);
-      char buffer[50];
-      auto size = std::strftime(buffer, sizeof(buffer), "%a, %d %b %Y %T GMT", &now);
-
-      rederedTime = oatpp::String(buffer, size, true);
-
-    }
-
-    return rederedTime;
-
-  }
-
   ENDPOINT_ASYNC("GET", "/plaintext", Plaintext) {
 
     ENDPOINT_ASYNC_INIT(Plaintext)
@@ -64,7 +38,7 @@ public:
     Action act() override {
       const auto& response = controller->createResponse(Status::CODE_200, oatpp::String("Hello, World!", 13, false));
       response->putHeader(Header::CONTENT_TYPE, oatpp::data::share::StringKeyLabel(nullptr, (p_char8)"text/plain", 10));
-      response->putHeader("Date", renderTime());
+      response->putHeader("Date", Utils::renderTime());
       return _return(response);
     }
 
@@ -78,7 +52,7 @@ public:
       auto dto = MessageDto::createShared();
       dto->message = oatpp::String("Hello, World!", 13, false);
       const auto& response = controller->createDtoResponse(Status::CODE_200, dto);
-      response->putHeader("Date", renderTime());
+      response->putHeader("Date", Utils::renderTime());
       return _return(response);
     }
